+++ ledger/simulation/tracer.go
@@ -130,13 +130,13 @@ func (tracer *evalTracer) BeforeTxnGroup(ep *logic.EvalParams) {
 	if ep.GetCaller() != nil {
 		// If this is an inner txn group, save the txns
 		tracer.populateInnerTransactions(ep.TxnGroup)
-		tracer.result.TxnGroups[0].BudgetAdded += ep.Proto.MaxAppProgramCost
+		tracer.result.TxnGroups[0].BudgetAdded += uint64(ep.Proto.MaxAppProgramCost)
 	}
 	tracer.cursorEvalTracer.BeforeTxnGroup(ep)
 
 	// Currently only supports one (first) txn group
 	if ep.PooledApplicationBudget != nil && tracer.result.TxnGroups[0].BudgetAdded == 0 {
-		tracer.result.TxnGroups[0].BudgetAdded = *ep.PooledApplicationBudget
+		tracer.result.TxnGroups[0].BudgetAdded = uint64(*ep.PooledApplicationBudget)
 	}
 }
 
@@ -197,7 +197,7 @@ func (tracer *evalTracer) AfterProgram(cx *logic.EvalContext, evalError error) {
 	}
 
 	// Report cost of this program
-	tracer.result.TxnGroups[0].Txns[tracer.relativeCursor[0]].BudgetUsed += cx.Cost()
+	tracer.result.TxnGroups[0].Txns[tracer.relativeCursor[0]].BudgetUsed += uint64(cx.Cost())
 
 	tracer.handleError(evalError)
 }
